creating global exception

+++ src/main/java/com/example/showroommanagement/service/ManagerService.java
@@ -1,13 +1,13 @@
 package com.example.showroommanagement.service;
 
+import com.example.showroommanagement.dto.ResponseDTO;
 import com.example.showroommanagement.entity.Manager;
+import com.example.showroommanagement.exception.BadRequestServiceAlertException;
 import com.example.showroommanagement.repository.ManagerRepository;
+import com.example.showroommanagement.util.Constant;
 import jakarta.transaction.Transactional;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
-
-import java.util.List;
-import java.util.Map;
-import java.util.NoSuchElementException;
 
 @Service
 public class ManagerService {
@@ -18,22 +18,40 @@ public class ManagerService {
         this.managerRepository = managerRepository;
     }
 
-    public Manager createManager(final Manager Manager) {
-        return this.managerRepository.save(Manager);
+    public ResponseDTO createManager(final Manager manager) {
+        final ResponseDTO responseDTO = new ResponseDTO();
+        responseDTO.setStatusCode(HttpStatus.OK.value());
+        responseDTO.setMessage(Constant.CREATE);
+        responseDTO.setData(this.managerRepository.save(manager));
+        return responseDTO;
     }
 
-    public Manager retrieveManagerById(final Integer id) {
-        return this.managerRepository.findById(id).orElse(null);
-
+    public ResponseDTO retrieveManagerById(final Integer id) {
+        if(this.managerRepository.existsById(id))
+        { this.managerRepository.findById(id);
+            final ResponseDTO responseDTO = new ResponseDTO();
+            responseDTO.setStatusCode(HttpStatus.OK.value());
+            responseDTO.setMessage(Constant.RETRIEVE);
+            responseDTO.setData(this.managerRepository.findById(id));
+            return responseDTO;
+        }
+         else {
+            throw new BadRequestServiceAlertException(Constant.ID_DOES_NOT_EXIST);
+        }
     }
 
-    public List<Manager> retrieveManager() {
-        return this.managerRepository.findAll();
+    public ResponseDTO retrieveManager() {
+        final ResponseDTO responseDTO = new ResponseDTO();
+        responseDTO.setMessage(Constant.RETRIEVE);
+        responseDTO.setStatusCode(HttpStatus.OK.value());
+        responseDTO.setData(this.managerRepository.findAll());
+        return responseDTO;
     }
+
 
     @Transactional
-    public Manager updateManagerById(final Manager manager, Integer id) {
-        final Manager existingManager = this.managerRepository.findById(id).orElseThrow(() -> new NoSuchElementException("Data not found"));
+    public ResponseDTO updateManagerById(final Manager manager, Integer id) {
+        final Manager existingManager = this.managerRepository.findById(id).orElseThrow(() -> new BadRequestServiceAlertException(Constant.ID_DOES_NOT_EXIST));
         if (manager.getName() != null) {
             existingManager.setName(manager.getName());
         }
@@ -49,14 +67,28 @@ public class ManagerService {
         if (manager.getPassword() != null) {
             existingManager.setPassword(manager.getPassword());
         }
-        return this.managerRepository.save(existingManager);
+        final ResponseDTO responseDTO = new ResponseDTO();
+        responseDTO.setMessage(Constant.UPDATE);
+        responseDTO.setStatusCode(HttpStatus.OK.value());
+        responseDTO.setData(this.managerRepository.save(existingManager));
+        return responseDTO;
     }
 
     @Transactional
-    public Map<String, String> removeManagerById(final Integer id) {
-        this.managerRepository.findById(id).orElseThrow(() -> new NoSuchElementException("Manager not found with id: " + id));
-        this.managerRepository.deleteById(id);
-        return (Map.of("message", "Manager deleted successfully."));
+    public ResponseDTO removeManagerById(final Integer id) {
+        if (id == null) {
+            throw new BadRequestServiceAlertException(Constant.DATA_NULL);
+        }
+        if (this.managerRepository.existsById(id)) {
+            this.managerRepository.deleteById(id);
+            final ResponseDTO responseDTO = new ResponseDTO();
+            responseDTO.setMessage(Constant.DELETE);
+            responseDTO.setStatusCode(HttpStatus.OK.value());
+            responseDTO.setData(Constant.REMOVE);
+            return responseDTO;
+        } else {
+            throw new BadRequestServiceAlertException(Constant.ID_DOES_NOT_EXIST);
+        }
     }
 }
