+++ src/main/java/com/example/showroommanagement/service/ProductService.java
@@ -1,13 +1,13 @@
 package com.example.showroommanagement.service;
 
+import com.example.showroommanagement.dto.ResponseDTO;
 import com.example.showroommanagement.entity.Product;
+import com.example.showroommanagement.exception.BadRequestServiceAlertException;
 import com.example.showroommanagement.repository.ProductRepository;
+import com.example.showroommanagement.util.Constant;
 import jakarta.transaction.Transactional;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
-
-import java.util.List;
-import java.util.Map;
-import java.util.NoSuchElementException;
 
 @Service
 public class ProductService {
@@ -17,22 +17,39 @@ public class ProductService {
         this.productRepository = productRepository;
     }
 
-    public Product createProduct(final Product product) {
-        return this.productRepository.save(product);
+    public ResponseDTO createProduct(final Product product) {
+        final ResponseDTO responseDTO = new ResponseDTO();
+        responseDTO.setStatusCode(HttpStatus.OK.value());
+        responseDTO.setData(this.productRepository.save(product));
+        responseDTO.setMessage(Constant.CREATE);
+        return responseDTO;
     }
 
-    public Product retrieveProductById(final Integer id) {
-        return this.productRepository.findById(id).orElse(null);
-
+    public ResponseDTO retrieveProductById(final Integer id) {
+        if (this.productRepository.existsById(id)) {
+            this.productRepository.findById(id);
+            final ResponseDTO responseDTO = new ResponseDTO();
+            responseDTO.setStatusCode(HttpStatus.OK.value());
+            responseDTO.setData(this.productRepository.findById(id));
+            responseDTO.setMessage(Constant.RETRIEVE);
+            return responseDTO;
+        } else {
+            throw new BadRequestServiceAlertException(Constant.ID_DOES_NOT_EXIST);
+        }
     }
 
-    public List<Product> retrieveProduct() {
-        return this.productRepository.findAll();
+
+    public ResponseDTO retrieveProduct() {
+        final ResponseDTO responseDTO = new ResponseDTO();
+        responseDTO.setMessage(Constant.RETRIEVE);
+        responseDTO.setStatusCode(HttpStatus.OK.value());
+        responseDTO.setData(this.productRepository.findAll());
+        return responseDTO;
     }
 
     @Transactional
-    public Product updateProductById(final Product product, Integer id) {
-        final Product existingProduct = this.productRepository.findById(id).orElseThrow(() -> new NoSuchElementException("Data not found"));
+    public ResponseDTO updateProductById(final Product product, Integer id) {
+        final Product existingProduct = this.productRepository.findById(id).orElseThrow(() -> new BadRequestServiceAlertException(Constant.ID_DOES_NOT_EXIST));
         if (product.getId() != null) {
             existingProduct.setId(product.getId());
         }
@@ -48,15 +65,28 @@ public class ProductService {
         if (product.getColour() != null) {
             existingProduct.setColour(product.getColour());
         }
-
-        return this.productRepository.save(existingProduct);
+        final ResponseDTO responseDTO = new ResponseDTO();
+        responseDTO.setMessage(Constant.UPDATE);
+        responseDTO.setStatusCode(HttpStatus.OK.value());
+        responseDTO.setData(this.productRepository.save(existingProduct));
+        return responseDTO;
     }
 
-    @Transactional
-    public Map<String, String> removeProductById(final Integer id) {
-        this.productRepository.findById(id).orElseThrow(() -> new NoSuchElementException("Product not found with id: " + id));
-        this.productRepository.deleteById(id);
-        return (Map.of("message", "Product deleted successfully."));
+    public ResponseDTO removeProductById(final Integer id) {
+        if (id == null) {
+            throw new BadRequestServiceAlertException(Constant.DATA_NULL);
+        }
+        if (this.productRepository.existsById(id)) {
+            this.productRepository.deleteById(id);
+            final ResponseDTO responseDTO = new ResponseDTO();
+            responseDTO.setMessage(Constant.DELETE);
+            responseDTO.setStatusCode(HttpStatus.OK.value());
+            responseDTO.setData(Constant.REMOVE);
+            return responseDTO;
+        } else {
+            throw new BadRequestServiceAlertException(Constant.ID_DOES_NOT_EXIST);
+        }
     }
-
 }
+
+
