+++ src/main/java/com/example/showroommanagement/service/SaleDetailsService.java
@@ -1,15 +1,16 @@
 package com.example.showroommanagement.service;
 
+import com.example.showroommanagement.dto.ResponseDTO;
 import com.example.showroommanagement.dto.SalesDetailsDTO;
 import com.example.showroommanagement.entity.SaleDetails;
+import com.example.showroommanagement.exception.BadRequestServiceAlertException;
 import com.example.showroommanagement.repository.SaleDetailsRepository;
+import com.example.showroommanagement.util.Constant;
 import jakarta.transaction.Transactional;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Map;
-import java.util.NoSuchElementException;
+import java.util.*;
 
 @Service
 public class SaleDetailsService {
@@ -19,23 +20,38 @@ public class SaleDetailsService {
         this.saleDetailsRepository = saleDetailsRepository;
     }
 
-    public SaleDetails createSales(final SaleDetails salesDetails) {
-
-        return this.saleDetailsRepository.save(salesDetails);
-    }
-
-    public SaleDetails retrieveSalesById(final Integer id) {
-        return this.saleDetailsRepository.findById(id).orElse(null);
-
-    }
-
-    public List<SaleDetails> retrieveSales() {
-        return this.saleDetailsRepository.findAll();
-    }
-
     @Transactional
-    public SaleDetails updateSalesById(final SaleDetails salesDetails, Integer id) {
-        final SaleDetails existingSalesDetails = this.saleDetailsRepository.findById(id).orElseThrow(() -> new NoSuchElementException("Data not found"));
+    public ResponseDTO createSales(final SaleDetails salesDetails) {
+        final ResponseDTO responseDTO = new ResponseDTO();
+        responseDTO.setMessage(Constant.CREATE);
+        responseDTO.setStatusCode(HttpStatus.CREATED.value());
+        responseDTO.setData(this.saleDetailsRepository.save(salesDetails));
+        return responseDTO;
+    }
+
+
+    public ResponseDTO retrieveSalesById(final Integer id) {
+        if (this.saleDetailsRepository.existsById(id)) {
+            final ResponseDTO responseDTO = new ResponseDTO();
+            responseDTO.setMessage(Constant.RETRIEVE);
+            responseDTO.setStatusCode(HttpStatus.OK.value());
+            responseDTO.setData(this.saleDetailsRepository.findById(id));
+            return responseDTO;
+        } else {
+            throw new BadRequestServiceAlertException(Constant.ID_DOES_NOT_EXIST);
+        }
+    }
+
+    public ResponseDTO retrieveSales() {
+        final ResponseDTO responseDTO = new ResponseDTO();
+        responseDTO.setMessage(Constant.RETRIEVE);
+        responseDTO.setStatusCode(HttpStatus.OK.value());
+        responseDTO.setData(this.saleDetailsRepository.findAll());
+        return responseDTO;
+    }
+    @Transactional
+    public ResponseDTO updateSalesById(final SaleDetails salesDetails, Integer id) {
+        final SaleDetails existingSalesDetails = this.saleDetailsRepository.findById(id).orElseThrow(() -> new BadRequestServiceAlertException(Constant.ID_DOES_NOT_EXIST));
         if (salesDetails.getId() != null) {
             existingSalesDetails.setId(salesDetails.getId());
         }
@@ -48,23 +64,38 @@ public class SaleDetailsService {
         if (salesDetails.getCustomer() != null) {
             existingSalesDetails.setCustomer(salesDetails.getCustomer());
         }
-        return this.saleDetailsRepository.save(existingSalesDetails);
+       final ResponseDTO responseDTO = new ResponseDTO();
+        responseDTO.setMessage(Constant.UPDATE);
+        responseDTO.setStatusCode(HttpStatus.OK.value());
+        responseDTO.setData(this.saleDetailsRepository.save(existingSalesDetails));
+        return responseDTO;
     }
 
     @Transactional
-    public Map<String, String> removeSalesById(final Integer id) {
-        this.saleDetailsRepository.findById(id).orElseThrow(() -> new NoSuchElementException("saledetails not found with id: " + id));
-        this.saleDetailsRepository.deleteById(id);
-        return (Map.of("message", "saledatails deleted successfully."));
+    public ResponseDTO removeSalesById(final Integer id) {
+        if (id == null) {
+            throw new BadRequestServiceAlertException(Constant.DATA_NULL);
+        }
+        if (this.saleDetailsRepository.existsById(id)) {
+            this.saleDetailsRepository.deleteById(id);
+            final ResponseDTO responseDTO = new ResponseDTO();
+            responseDTO.setMessage(Constant.DELETE);
+            responseDTO.setStatusCode(HttpStatus.OK.value());
+            responseDTO.setData(Constant.REMOVE);
+            return responseDTO;
+        } else {
+            throw new BadRequestServiceAlertException(Constant.ID_DOES_NOT_EXIST);
+        }
     }
 
     @Transactional
-    public List<SalesDetailsDTO> retrieveSalesDetails() {
+    public ResponseDTO retrieveSalesDetails(final String showroomName, final String bikeName) {
         List<SaleDetails> retrieveSales = this.saleDetailsRepository.findAll();
         List<SalesDetailsDTO> salesDetailsDTOS = new ArrayList<>();
         for (SaleDetails salesDetails : retrieveSales) {
             SalesDetailsDTO salesDetailsDTO = new SalesDetailsDTO();
             salesDetailsDTO.setShowroomName(salesDetails.getProduct().getSalesMan().getShowroom().getName());
+            salesDetailsDTO.setShowroomBrand(salesDetails.getProduct().getSalesMan().getShowroom().getBrand());
             salesDetailsDTO.setProductModel(salesDetails.getProduct().getModel());
             salesDetailsDTO.setProductprice(salesDetails.getProduct().getPrice());
             salesDetailsDTO.setSalesmanName(salesDetails.getProduct().getSalesMan().getName());
@@ -74,6 +105,12 @@ public class SaleDetailsService {
             salesDetailsDTO.setCustomerAddress(salesDetails.getCustomer().getAddress());
             salesDetailsDTOS.add(salesDetailsDTO);
         }
-        return salesDetailsDTOS;
+        final ResponseDTO responseDTO = new ResponseDTO();
+        responseDTO.setMessage(Constant.RETRIEVE);
+        responseDTO.setStatusCode(HttpStatus.OK.value());
+        responseDTO.setData( salesDetailsDTOS);
+        return responseDTO;
+
     }
+
 }
