+++ src/main/java/com/example/showroommanagement/service/SalesManService.java
@@ -1,15 +1,17 @@
 package com.example.showroommanagement.service;
 
+import com.example.showroommanagement.dto.ResponseDTO;
 import com.example.showroommanagement.dto.SalesManDetailsDTO;
 import com.example.showroommanagement.entity.SalesMan;
+import com.example.showroommanagement.exception.BadRequestServiceAlertException;
 import com.example.showroommanagement.repository.SalesManRepository;
+import com.example.showroommanagement.util.Constant;
 import jakarta.transaction.Transactional;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Map;
-import java.util.NoSuchElementException;
 
 @Service
 public class SalesManService {
@@ -20,22 +22,39 @@ public class SalesManService {
         this.salesManRepository = salesManRepository;
     }
 
-    public SalesMan createSalesMan(final SalesMan salesMan) {
-        return this.salesManRepository.save(salesMan);
+    @Transactional
+    public ResponseDTO createSalesMan(final SalesMan salesman) {
+        final ResponseDTO responseDTO = new ResponseDTO();
+        responseDTO.setMessage(Constant.CREATE);
+        responseDTO.setStatusCode(HttpStatus.CREATED.value());
+        responseDTO.setData(this.salesManRepository.save(salesman));
+        return responseDTO;
     }
 
-    public SalesMan retrieveSalesManById(final Integer id) {
-        return this.salesManRepository.findById(id).orElse(null);
-
+    public ResponseDTO retrieveSalesManById(final Integer id) {
+        if (this.salesManRepository.existsById(id)) {
+            this.salesManRepository.findById(id);
+            final ResponseDTO responseDTO = new ResponseDTO();
+            responseDTO.setMessage(Constant.RETRIEVE);
+            responseDTO.setStatusCode(HttpStatus.OK.value());
+            responseDTO.setData(this.salesManRepository.findById(id));
+            return responseDTO;
+        } else {
+            throw new BadRequestServiceAlertException(Constant.ID_DOES_NOT_EXIST);
+        }
     }
 
-    public List<SalesMan> retrieveSalesMan() {
-        return this.salesManRepository.findAll();
+    public ResponseDTO retrieveSalesMan() {
+        final ResponseDTO responseDTO = new ResponseDTO();
+        responseDTO.setMessage(Constant.RETRIEVE);
+        responseDTO.setStatusCode(HttpStatus.OK.value());
+        responseDTO.setData(this.salesManRepository.findAll());
+        return responseDTO;
     }
 
     @Transactional
-    public SalesMan updateSalesManById(final SalesMan salesMan, Integer id) {
-        final SalesMan existingSalesMan = this.salesManRepository.findById(id).orElseThrow(() -> new NoSuchElementException("Data not found"));
+    public ResponseDTO updateSalesManById(final SalesMan salesMan, Integer id) {
+        final SalesMan existingSalesMan = this.salesManRepository.findById(id).orElseThrow(() -> new BadRequestServiceAlertException(Constant.ID_DOES_NOT_EXIST));
         if (salesMan.getId() != null) {
             existingSalesMan.setId(salesMan.getId());
         }
@@ -51,14 +70,30 @@ public class SalesManService {
         if (salesMan.getShowroom() != null) {
             existingSalesMan.setShowroom(salesMan.getShowroom());
         }
-        return this.salesManRepository.save(existingSalesMan);
+
+        final ResponseDTO responseDTO = new ResponseDTO();
+        responseDTO.setMessage(Constant.UPDATE);
+        responseDTO.setStatusCode(HttpStatus.OK.value());
+        responseDTO.setData(this.salesManRepository.save(existingSalesMan));
+        return responseDTO;
+
     }
 
     @Transactional
-    public Map<String, String> removeSalesManById(final Integer id) {
-        this.salesManRepository.findById(id).orElseThrow(() -> new NoSuchElementException("salesman not found with id: " + id));
-        this.salesManRepository.deleteById(id);
-        return (Map.of("message", "salesman deleted successfully."));
+    public ResponseDTO removeSalesManById(final Integer id) {
+        if (id == null) {
+            throw new BadRequestServiceAlertException(Constant.DATA_NULL);
+        }
+        if (this.salesManRepository.existsById(id)) {
+            this.salesManRepository.deleteById(id);
+            final ResponseDTO responseDTO = new ResponseDTO();
+            responseDTO.setMessage(Constant.DELETE);
+            responseDTO.setStatusCode(HttpStatus.OK.value());
+            responseDTO.setData(Constant.REMOVE);
+            return responseDTO;
+        } else {
+            throw new BadRequestServiceAlertException(Constant.ID_DOES_NOT_EXIST);
+        }
     }
